+++ geopackage-sdk/src/androidTest/java/mil/nga/geopackage/test/extension/elevation/ElevationTilesPngTestUtils.java
@@ -243,9 +243,9 @@ public class ElevationTilesPngTestUtils {
                 griddedTile.getTableName());
         long tableId = griddedTile.getTableId();
         TestCase.assertTrue(tableId >= 0);
-        TestCase.assertTrue(griddedTile.getScaleOrDefault() >= 0);
+        TestCase.assertTrue(griddedTile.getScale() >= 0);
         if (elevationTileValues != null) {
-            TestCase.assertTrue(griddedTile.getOffsetOrDefault() >= 0);
+            TestCase.assertTrue(griddedTile.getOffset() >= 0);
         }
         griddedTile.getMin();
         griddedTile.getMax();
@@ -309,10 +309,10 @@ public class ElevationTilesPngTestUtils {
                     TestCase.assertNull(elevationValue);
                 } else {
                     TestCase.assertEquals(
-                            (pixelValue * griddedTile.getScaleOrDefault() + griddedTile
-                                    .getOffsetOrDefault())
-                                    * griddedCoverage.getScaleOrDefault()
-                                    + griddedCoverage.getOffsetOrDefault(),
+                            (pixelValue * griddedTile.getScale() + griddedTile
+                                    .getOffset())
+                                    * griddedCoverage.getScale()
+                                    + griddedCoverage.getOffset(),
                             elevationValue);
                 }
             }