+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Adheres to [Semantic Versioning](http://semver.org/).
 
 * geopackage-core version updated to 1.2.2
 * tiff version updated to 1.0.1
+* Elevation Extension scale and offset columns changed to be non nullable
 * Android support library updated to 25.1.1
 * min SDK lowered from 14 to 13
 * Android Manifest cleanup